# openapi.yaml
openapi: 3.0.3
info:
  title: Booking API
  description: REST API for time-slot bookings with JWT auth and role-based access (Node.js + Express + SQLite)
  version: 1.0.0
servers:
  - url: /
paths:
  /api/auth/register:
    post:
      summary: Register a new user (first user becomes admin)
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, example: user@example.com }
                password: { type: string, example: secret123 }
      responses:
        '201': { description: Registered, returns JWT }
        '409': { description: Email already registered }
  /api/auth/login:
    post:
      summary: Login and get a JWT
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string }
                password: { type: string }
      responses:
        '200': { description: OK, returns JWT }
        '401': { description: Invalid credentials }
  /api/bookings:
    get:
      summary: List bookings (own for users, all for admin)
      tags: [Bookings]
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: List of bookings }
    post:
      summary: Create a booking (rejects overlapping slots)
      tags: [Bookings]
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title, starts_at, ends_at]
              properties:
                title: { type: string, example: Haircut }
                starts_at: { type: string, example: '2026-07-10T10:00:00Z' }
                ends_at: { type: string, example: '2026-07-10T11:00:00Z' }
      responses:
        '201': { description: Created }
        '409': { description: Slot overlaps existing booking }
  /api/bookings/{id}:
    get:
      summary: Get a booking by id
      tags: [Bookings]
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        '200': { description: Booking }
        '404': { description: Not found }
    delete:
      summary: Cancel a booking (soft delete)
      tags: [Bookings]
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        '200': { description: Cancelled }
  /api/bookings/{id}/hard:
    delete:
      summary: Permanently delete a booking (admin only)
      tags: [Bookings]
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        '204': { description: Deleted }
        '403': { description: Admin only }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT